+++ lib/src/panel.dart
@@ -1,6 +1,6 @@
 /*
 Name: Akshath Jain
-Date: 3/18/19
+Date: 3/18/19 - 4/2/19
 Purpose: Defines the sliding_up_panel widget
 Copyright: © 2019, Akshath Jain. All rights reserved.
 Licensing: More information can be found here: https://github.com/akshathjain/sliding_up_panel/blob/master/LICENSE
@@ -75,6 +75,24 @@ class SlidingUpPanel extends StatefulWidget {
   /// and 1.0 is completely opaque.
   final double backdropOpacity;
 
+  /// Flag that indicates whether or not tapping the
+  /// backdrop closes the panel. Defaults to true.
+  final bool backdropTapClosesPanel;
+
+  /// If non-null, this callback
+  /// is called as the panel slides around with the
+  /// current position of the panel. The position is a double
+  /// between 0.0 and 1.0 where 0.0 is fully collapsed and 1.0 is fully open.
+  final void Function(double position) onPanelSlide;
+
+  /// If non-null, this callback is called when the
+  /// panel is fully opened
+  final VoidCallback onPanelOpened;
+
+  /// If non-null, this callback is called when the panel
+  /// is fully collapsed.
+  final VoidCallback onPanelClosed;
+
   SlidingUpPanel({
     Key key,
     @required this.panel,
@@ -99,6 +117,10 @@ class SlidingUpPanel extends StatefulWidget {
     this.backdropEnabled = false,
     this.backdropColor = Colors.black,
     this.backdropOpacity = 0.5,
+    this.backdropTapClosesPanel = true,
+    this.onPanelSlide,
+    this.onPanelOpened,
+    this.onPanelClosed
   }) : assert(0 <= backdropOpacity && backdropOpacity <= 1.0),
        super(key: key);
 
@@ -110,10 +132,7 @@ class _SlidingUpPanelState extends State<SlidingUpPanel> with SingleTickerProvid
 
   AnimationController _ac;
 
-  double _closedHeight; //this can change depending on whether or not the user hides the sliding panel
-  double _openHeight; //this can change depending on whether or not the user hides the sliding panel
-
-  bool _isVisible = true;
+  bool _isPanelVisible = true;
 
   @override
   void initState(){
@@ -124,17 +143,27 @@ class _SlidingUpPanelState extends State<SlidingUpPanel> with SingleTickerProvid
       duration: const Duration(milliseconds: 300),
     )..addListener((){
       setState((){});
+
+      if(widget.onPanelSlide != null) widget.onPanelSlide(_ac.value);
+
+      if(widget.onPanelOpened != null && _ac.value == 1.0) widget.onPanelOpened();
+
+      if(widget.onPanelClosed != null && _ac.value == 0.0) widget.onPanelClosed();
     });
-    _ac.value = 0.0;
 
-    _closedHeight = widget.minHeight;
-    _openHeight = widget.maxHeight;
-
-    widget.controller?._addCloseListener(_close);
-    widget.controller?._addOpenListener(_open);
-    widget.controller?._addHideListener(_hide);
-    widget.controller?._addShowListener(_show);
-
+    widget.controller?._addListeners(
+      _close,
+      _open,
+      _hide,
+      _show,
+      _setPanelPosition,
+      _animatePanelToPosition,
+      _getPanelPosition,
+      _isPanelAnimating,
+      _isPanelOpen,
+      _isPanelClosed,
+      _isPanelShown,
+    );
   }
 
   @override
@@ -153,26 +182,29 @@ class _SlidingUpPanelState extends State<SlidingUpPanel> with SingleTickerProvid
 
 
         //the backdrop to overlay on the body
-        !widget.backdropEnabled ? Container() : Opacity(
-          opacity: _ac.value * widget.backdropOpacity,
-          child: Container(
-            height: MediaQuery.of(context).size.height,
-            width: MediaQuery.of(context).size.width,
+        !widget.backdropEnabled ? Container() : GestureDetector(
+          onTap: widget.backdropTapClosesPanel ? _close : null,
+          child: Opacity(
+            opacity: _ac.value * widget.backdropOpacity,
+            child: Container(
+              height: MediaQuery.of(context).size.height,
+              width: MediaQuery.of(context).size.width,
 
-            //set color to null so that touch events pass through
-            //to the body when the panel is closed, otherwise,
-            //if a color exists, then touch events won't go through
-            color: _ac.value == 0.0 ? null : widget.backdropColor,
+              //set color to null so that touch events pass through
+              //to the body when the panel is closed, otherwise,
+              //if a color exists, then touch events won't go through
+              color: _ac.value == 0.0 ? null : widget.backdropColor,
+            ),
           ),
         ),
 
 
         //the actual sliding part
-        !_isVisible ? Container() : GestureDetector(
+        !_isPanelVisible ? Container() : GestureDetector(
           onVerticalDragUpdate: _onDrag,
           onVerticalDragEnd: _onDragEnd,
           child: Container(
-            height: _ac.value * (_openHeight - _closedHeight) + _closedHeight,
+            height: _ac.value * (widget.maxHeight - widget.minHeight) + widget.minHeight,
             margin: widget.margin,
             padding: widget.padding,
             decoration: widget.renderPanelSheet ? BoxDecoration(
@@ -201,7 +233,13 @@ class _SlidingUpPanelState extends State<SlidingUpPanel> with SingleTickerProvid
                   height: widget.minHeight,
                   child: Opacity(
                     opacity: 1.0 - _ac.value,
-                    child: widget.collapsed ?? Container()
+
+                    // if the panel is open ignore pointers (touch events) on the collapsed
+                    // child so that way touch events go through to whatever is underneath
+                    child: IgnorePointer(
+                      ignoring: _isPanelOpen(),
+                      child: widget.collapsed ?? Container(),
+                    ),
                   ),
                 ),
 
@@ -222,7 +260,7 @@ class _SlidingUpPanelState extends State<SlidingUpPanel> with SingleTickerProvid
   }
 
   void _onDrag(DragUpdateDetails details){
-    _ac.value -= details.primaryDelta / (_openHeight - _closedHeight);
+    _ac.value -= details.primaryDelta / (widget.maxHeight - widget.minHeight);
   }
 
   void _onDragEnd(DragEndDetails details){
@@ -233,7 +271,7 @@ class _SlidingUpPanelState extends State<SlidingUpPanel> with SingleTickerProvid
 
     //check if the velocity is sufficient to constitute fling
     if(details.velocity.pixelsPerSecond.dy.abs() >= minFlingVelocity){
-      double visualVelocity = - details.velocity.pixelsPerSecond.dy / (_openHeight - _closedHeight);
+      double visualVelocity = - details.velocity.pixelsPerSecond.dy / (widget.maxHeight - widget.minHeight);
 
       if(widget.panelSnapping)
         _ac.fling(velocity: visualVelocity);
@@ -259,6 +297,13 @@ class _SlidingUpPanelState extends State<SlidingUpPanel> with SingleTickerProvid
 
   }
 
+
+
+
+  //---------------------------------
+  //PanelController related functions
+  //---------------------------------
+
   //close the panel
   void _close(){
     _ac.fling(velocity: -1.0);
@@ -273,7 +318,7 @@ class _SlidingUpPanelState extends State<SlidingUpPanel> with SingleTickerProvid
   void _hide(){
     _ac.fling(velocity: -1.0).then((x){
       setState(() {
-        _isVisible = false;
+        _isPanelVisible = false;
       });
     });
   }
@@ -282,48 +327,57 @@ class _SlidingUpPanelState extends State<SlidingUpPanel> with SingleTickerProvid
   void _show(){
     _ac.fling(velocity: -1.0).then((x){
       setState(() {
-        _isVisible = true;
+        _isPanelVisible = true;
       });
     });
   }
 
+  //set the panel position to value - must
+  //be between 0.0 and 1.0
+  void _setPanelPosition(double value){
+    assert(0.0 <= value && value <= 1.0);
+    _ac.value = value;
+  }
+
+  //set the panel position to value - must
+  //be between 0.0 and 1.0
+  void _animatePanelToPosition(double value){
+    assert(0.0 <= value && value <= 1.0);
+    _ac.animateTo(value);
+  }
+
+  //get the current panel position
+  //returns the % offset from collapsed state
+  //as a decimal between 0.0 and 1.0
+  double _getPanelPosition(){
+    return _ac.value;
+  }
+
+  //returns whether or not
+  //the panel is still animating
+  bool _isPanelAnimating(){
+    return _ac.isAnimating;
+  }
+
+  //returns whether or not the
+  //panel is open
+  bool _isPanelOpen(){
+    return _ac.value == 1.0;
+  }
+
+  //returns whether or not the
+  //panel is closed
+  bool _isPanelClosed(){
+    return _ac.value == 0.0;
+  }
+
+  //returns whether or not the
+  //panel is shown/hidden
+  bool _isPanelShown(){
+    return _isPanelVisible;
+  }
+
 }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 
 
 
@@ -337,9 +391,38 @@ class PanelController{
   VoidCallback _openListener;
   VoidCallback _hideListener;
   VoidCallback _showListener;
+  Function(double value) _setPanelPositionListener;
+  Function(double value) _setAnimatePanelToPositionListener;
+  double Function() _getPanelPositionListener;
+  bool Function() _isPanelAnimatingListener;
+  bool Function() _isPanelOpenListener;
+  bool Function() _isPanelClosedListener;
+  bool Function() _isPanelShownListener;
 
-  void _addCloseListener(VoidCallback listener){
-    this._closeListener = listener;
+  void _addListeners(
+    VoidCallback closeListener,
+    VoidCallback openListener,
+    VoidCallback hideListener,
+    VoidCallback showListener,
+    Function(double value) setPanelPositionListener,
+    Function(double value) setAnimatePanelToPositionListener,
+    double Function() getPanelPositionListener,
+    bool Function() isPanelAnimatingListener,
+    bool Function() isPanelOpenListener,
+    bool Function() isPanelClosedListener,
+    bool Function() isPanelShownListener,
+  ){
+    this._closeListener = closeListener;
+    this._openListener = openListener;
+    this._hideListener = hideListener;
+    this._showListener = showListener;
+    this._setPanelPositionListener = setPanelPositionListener;
+    this._setAnimatePanelToPositionListener = setAnimatePanelToPositionListener;
+    this._getPanelPositionListener = getPanelPositionListener;
+    this._isPanelAnimatingListener = isPanelAnimatingListener;
+    this._isPanelOpenListener = isPanelOpenListener;
+    this._isPanelClosedListener = isPanelClosedListener;
+    this._isPanelShownListener = isPanelShownListener;
   }
 
   /// Closes the sliding panel to its collapsed state (i.e. to the  minHeight)
@@ -347,17 +430,10 @@ class PanelController{
     _closeListener();
   }
 
-  void _addOpenListener(VoidCallback listener){
-    this._openListener = listener;
-  }
-
-  /// Opens the sliding panel fully (i.e. to the maxHeight)
+  /// Opens the sliding panel fully
+  /// (i.e. to the maxHeight)
   void open(){
     _openListener();
-  }
-
-  void _addHideListener(VoidCallback listener){
-    this._hideListener = listener;
   }
 
   /// Hides the sliding panel (i.e. is invisible)
@@ -365,13 +441,60 @@ class PanelController{
     _hideListener();
   }
 
-  void _addShowListener(VoidCallback listener){
-    this._showListener = listener;
-  }
-
-  /// Shows the sliding panel in its collapsed state (i.e. "un-hide" the sliding panel)
+  /// Shows the sliding panel in its collapsed state
+  /// (i.e. "un-hide" the sliding panel)
   void show(){
     _showListener();
+  }
+
+  /// Sets the panel position (without animation).
+  /// The value must between 0.0 and 1.0
+  /// where 0.0 is fully collapsed and 1.0 is completely open.
+  void setPanelPosition(double value){
+    assert(0.0 <= value && value <= 1.0);
+    _setPanelPositionListener(value);
+  }
+
+  /// Animates the panel position to the value.
+  /// The value must between 0.0 and 1.0
+  /// where 0.0 is fully collapsed and 1.0 is completely open
+  void animatePanelToPosition(double value){
+    assert(0.0 <= value && value <= 1.0);
+    _setAnimatePanelToPositionListener(value);
+  }
+
+  /// Gets the current panel position.
+  /// Returns the % offset from collapsed state
+  /// to the open state
+  /// as a decimal between 0.0 and 1.0
+  /// where 0.0 is fully collapsed and
+  /// 1.0 is full open.
+  double getPanelPosition(){
+    return _getPanelPositionListener();
+  }
+
+  /// Returns whether or not the panel is
+  /// currently animating.
+  bool isPanelAnimating(){
+    return _isPanelAnimatingListener();
+  }
+
+  /// Returns whether or not the
+  /// panel is open.
+  bool isPanelOpen(){
+    return _isPanelOpenListener();
+  }
+
+  /// Returns whether or not the
+  /// panel is closed.
+  bool isPanelClosed(){
+    return _isPanelClosedListener();
+  }
+
+  /// Returns whether or not the
+  /// panel is shown/hidden.
+  bool isPanelShown(){
+    return _isPanelShownListener();
   }
 
 }